+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,9 @@
         android:theme="@style/AppTheme">
 
         <activity android:name=".activities.Activator"/>
-        <activity android:name=".activities.FragmentsActivity"/>
+        <activity android:name=".activities.MainActivity"/>
         <activity android:name=".activities.MovieInfoActivity"/>
-        <activity android:name=".activities.MainActivity">
+        <activity android:name=".activities.SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
