+++ app/src/main/java/com/example/movie_db/AdapterForPager.kt
